Email Reply Automation/requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
python-dotenv==1.0.0
google-generativeai==0.3.0 